pyproject.toml
[project]
name = "health-claim-checker"
version = "0.1.0"
description = "Multi-agent system for verifying health claims against peer-reviewed research"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    # Core
    "langchain>=0.1.0",
    "langgraph>=0.0.20",
    "anthropic>=0.18.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    # Retrieval
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",
    "biopython>=1.80",
    # Medical NLP
    "scispacy>=0.5.0",
    # Data
    "pandas>=2.0.0",
    "pdfplumber>=0.9.0",
    "requests>=2.28.0",
    # UI
    "streamlit>=1.28.0",
    # Observability
    "langfuse>=2.0.0",
    # Dev
    "pytest>=7.0.0",
    "python-dotenv>=1.0.0",
    "ipykernel>=6.0.0",
    "snowballstemmer>=3.0.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
